day 22: half working, half manual hacks
I wanted to implement a full solve of this with any input, but ran out
of patience after getting the cube movement on the "standardized" cube
so there is a lot of code that is specific to the input I had.

The idea was to transform the input into a standard cube so that all of
the movements would be the same. The code that transforms my input into
the standard cube is manually configured to work.

Standard Cube (using Rubix Cube terms):

    L U R
      F
      D
      B

Input Cube:

      U R
      F
    L D
    B

To transform this cube you need to move L next to U (which requires a
180 rotation) and B below D (-90 rotation).

After the input is transformed into the standard cube the moves are ran
against that cube. From there you get a final position, BUT that final
position COULD be in one of the sections that was moved to get to a
standard cube. Luckily that did not happen in this input! :) If it did,
the process to find the final position would be to reverse the rotations
on the section (tracking the final coordinate) and then translate that
coordinate by the amount that the section was moved to assemble the
standard cube (i.e. if the coord was in B you would do a 90 rotation
then -50 from the column).

Anyway, I left these notes in case I want to come back and finish this
the "correct" way.