day 16: even faster w/o caching
Instead of caching the intermediate states, generate all of the states
and use that to determine which states do not overlap for part 2. This
is much faster than running through the simulation for each possible
state pair (440ms vs ~3s).

This approach is much slower for part 1 (~125ms vs ~10ms), but that's
still within acceptable range; to get the optimal solution to both you
could keep the old cached solution for part 1.

Inspired by:
https://www.reddit.com/r/adventofcode/comments/zn6k1l/comment/j0hrdpt